Move update_database functions into database_helpers.
